# prop/portal/openapi/prop-api.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "Prop.Api | v1",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/portal/orders/{orderId}": {
      "get": {
        "tags": [
          "Portal"
        ],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "token",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BuyerOrderResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "BuyerOrderResponse": {
        "required": [
          "id",
          "number",
          "status",
          "email",
          "challengeId",
          "challengeName",
          "amount",
          "currency",
          "provider",
          "checkoutUrl",
          "accountId",
          "problem",
          "canLogIn",
          "inviteSentAt",
          "canChoosePassword"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "number": {
            "type": "integer",
            "format": "int64"
          },
          "status": {
            "$ref": "#/components/schemas/OrderStatus"
          },
          "email": {
            "type": "string"
          },
          "challengeId": {
            "type": "string"
          },
          "challengeName": {
            "type": "string"
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "currency": {
            "type": "string"
          },
          "provider": {
            "$ref": "#/components/schemas/PaymentProvider"
          },
          "checkoutUrl": {
            "type": [
              "null",
              "string"
            ],
            "format": "uri"
          },
          "accountId": {
            "type": [
              "null",
              "string"
            ],
            "format": "uuid"
          },
          "problem": {
            "type": [
              "null",
              "string"
            ]
          },
          "canLogIn": {
            "type": "boolean"
          },
          "inviteSentAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "canChoosePassword": {
            "type": "boolean"
          },
          "discountCode": {
            "type": [
              "null",
              "string"
            ]
          },
          "listAmount": {
            "type": [
              "null",
              "number"
            ],
            "format": "double"
          }
        },
        "description": "The order as its buyer sees it, with the price before its DiscountCode as ListAmount.\nCheckoutUrl is set while it waits for payment.\nCanLogIn means the trader already has a password for the portal.\nInviteSentAt is when the platform last emailed an invitation to choose one.\nCanChoosePassword means the buyer can choose the password right on the order's page: the order\nstarted the trader's only account, and the trader has no password yet."
      },
...
      "OrderStatus": {
        "enum": [
          "Pending",
          "Paid",
          "Expired"
        ],
        "description": "Where a purchase in the portal is. A pending order past its time counts as expired, but a payment that arrives later still counts."
      },
...
      "PaymentProvider": {
        "enum": [
          "Test",
          "Stripe",
          "External"
        ],
        "description": "Who takes the trader's money when a challenge is bought in the firm's portal (ADR 0019)."
      },
...
    }
  },
...
}