# prop/portal/src/lib/api/types.ts
import type { components } from "./schema";

type Schemas = components["schemas"];

export type Account = Schemas["AccountResponse"];
export type AccountDetails = Schemas["AccountDetailsResponse"];
export type LiveFigures = Schemas["LiveFigures"];
export type StageSummary = Schemas["StageResponse"];
export type StageProgress = Schemas["StageProgress"];
export type Results = Schemas["ResultsResponse"];
export type Performance = Schemas["PerformanceResponse"];
export type BalancePoint = Schemas["BalancePoint"];
export type BalanceChangeKind = Schemas["BalanceChangeKind"];
export type FloorPoint = Schemas["FloorPoint"];
export type DayResult = Schemas["DayResultResponse"];
export type TradeStatistics = Schemas["TradeStatisticsResponse"];
export type Trades = Schemas["TradesResponse"];
export type Trade = Schemas["TradeResponse"];
export type TradeSide = Schemas["TradeSide"];
export type TraderPayouts = Schemas["TraderPayoutsResponse"];
export type PayoutTotal = Schemas["PayoutTotalResponse"];
export type FloorFigure = Schemas["FloorFigure"];
export type BreachEvidence = Schemas["BreachEvidence"];
export type ExpiryEvidence = Schemas["ExpiryEvidence"];
export type ExpiryReason = Schemas["ExpiryReason"];
export type Branding = Schemas["BrandingResponse"];
export type ChallengeDefinition = Schemas["ChallengeDefinition"];
export type ChallengeStatus = Schemas["ChallengeStatus"];
export type FailureReason = Schemas["FailureReason"];
export type Me = Schemas["PortalMeResponse"];
export type Step = Schemas["StepResponse"];
export type Payout = Schemas["PayoutResponse"];
export type PayoutMethod = Schemas["PayoutMethod"];
export type PayoutMethodKind = Schemas["PayoutMethodKind"];
export type PayoutQuote = Schemas["PayoutQuoteResponse"];
export type PayoutStatus = Schemas["PayoutStatus"];
export type FirmStatus = Schemas["FirmStatus"];
export type FirmSettings = Schemas["FirmSettingsResponse"];
export type Platform = Schemas["PlatformResponse"];
export type Admins = Schemas["AdminsResponse"];
export type ChallengeTemplate = Schemas["ChallengeTemplateResponse"];
export type StageRules = Schemas["StageRules"];
export type Shop = Schemas["ShopResponse"];
export type ShopItem = Schemas["ShopItemResponse"];
export type Order = Schemas["OrderResponse"];
export type OrderDetails = Schemas["OrderDetailsResponse"];
export type OrderStatus = Schemas["OrderStatus"];
export type BuyerOrder = Schemas["BuyerOrderResponse"];
export type PaymentProvider = Schemas["PaymentProvider"];
export type PaymentSettings = Schemas["PaymentSettingsResponse"];
export type ChallengePrice = Schemas["ChallengePrice"];
export type Billing = Schemas["BillingResponse"];
export type Slots = Schemas["SlotsResponse"];
export type Charge = Schemas["ChargeResponse"];
export type ChargeLine = Schemas["ChargeLine"];
export type ChargeKind = Schemas["ChargeKind"];
export type ChargeStatus = Schemas["ChargeStatus"];
export type Card = Schemas["CardResponse"];
export type Prices = Schemas["PricesResponse"];
export type Quote = Schemas["QuoteResponse"];
export type Vat = Schemas["VatResponse"];
export type VatTreatment = Schemas["VatTreatment"];
export type FieldProblem = Schemas["FieldProblem"];
export type TestBillingCheckout = Schemas["TestCheckoutResponse"];
export type Suspension = Schemas["SuspensionResponse"];
export type VerificationResponse = Schemas["VerificationResponse"];
// The review's status is never null here, though the schema shares it with places where a firm has no review.
export type Verification = Omit<VerificationResponse, "status"> & { status: ReviewStatus };
export type FirmApplication = Schemas["FirmApplication"];
export type FirmOwner = Schemas["FirmOwner"];
export type FirmDocument = Schemas["DocumentResponse"];
// Nullable where a firm has no review, so the status itself is taken without null.
export type ReviewStatus = NonNullable<Schemas["ReviewStatus"]>;
export type OpsSite = Schemas["OpsSiteResponse"];
export type OpsMe = Schemas["OpsMeResponse"];
export type OpsFirm = Schemas["OpsFirmResponse"];
export type OpsEvent = Schemas["OpsEventResponse"];
export type OpsFirmGroup = Schemas["OpsFirmGroup"];
export type OpsFirmStage = Schemas["OpsFirmStage"];
export type OpsFirmCounts = Schemas["OpsFirmCountsResponse"];
export type OpsFirms = Schemas["OpsFirmsResponse"];
export type OpsFirmListItem = Schemas["OpsFirmListItemResponse"];
export type OpsOverview = Schemas["OpsOverviewResponse"];
export type OpsNeedsUs = Schemas["OpsNeedsUsResponse"];
export type OpsCharge = Schemas["OpsChargeResponse"];
export type OpsLatePayouts = Schemas["OpsLatePayoutsResponse"];
export type OpsWeek = Schemas["OpsWeekResponse"];
export type OpsFunnel = Schemas["OpsFunnelResponse"];
export type OpsActivity = Schemas["OpsActivityResponse"];
export type OpsActivityKind = Schemas["OpsActivityKind"];
export type OpsWaiting = Schemas["OpsWaitingResponse"];
export type OpsBillingOverview = Schemas["OpsBillingOverviewResponse"];
export type OpsNextMonthFirm = Schemas["OpsNextMonthFirmResponse"];
export type OpsPayouts = Schemas["OpsPayoutsResponse"];
export type OpsWaitingPayout = Schemas["OpsWaitingPayoutResponse"];
export type ReviewCheck = Schemas["ReviewCheckResponse"];
export type AccountGroup = Schemas["AccountGroup"];
export type AccountCounts = Schemas["AccountCountsResponse"];
export type AdminAccounts = Schemas["AdminAccountsResponse"];
export type MoneyTotal = Schemas["MoneyTotalResponse"];
export type PayoutGroup = Schemas["PayoutGroupResponse"];
export type PayoutSummary = Schemas["PayoutSummaryResponse"];
export type AdminPayout = Schemas["AdminPayoutResponse"];
export type AdminOverview = Schemas["AdminOverviewResponse"];
export type Week = Schemas["WeekResponse"];
export type Activity = Schemas["ActivityResponse"];
export type ActivityKind = Schemas["ActivityKind"];
export type PassRate = Schemas["PassRateResponse"];
export type ChallengeFigures = Schemas["ChallengeFiguresResponse"];
export type TraderSummary = Schemas["TraderSummaryResponse"];
export type TraderEmailKind = Schemas["TraderEmailKind"];
export type LinkCheck = Schemas["LinkCheckResponse"];
export type WebhookOverview = Schemas["WebhookOverviewResponse"];
export type WebhookDelivery = Schemas["WebhookDeliveryResponse"];
export type TradingConditions = Schemas["TradingConditionsResponse"];
export type TradingConditionsSymbol = Schemas["TradingConditionsSymbol"];
export type TradingSymbolRequest = Schemas["TradingSymbolRequest"];
export type DiscountCode = Schemas["DiscountCodeResponse"];
export type DiscountQuote = Schemas["DiscountQuoteResponse"];
export type TraderCheck = Schemas["TraderCheckResponse"];
export type RetryOffer = Schemas["RetryOffer"];
export type BreachClose = Schemas["BreachClose"];
export type CustomDomain = Schemas["DomainResponse"];